+++ include/gcli/github/status.h
@@ -36,8 +36,7 @@
 
 #include <gcli/status.h>
 
-size_t github_get_notifications(gcli_notification **notifications,
-                                int count);
+int github_get_notifications(gcli_notification **notifications, int count);
 
 int github_notification_mark_as_read(char const *id);
