+++ include/gcli/gitlab/status.h
@@ -36,8 +36,7 @@
 
 #include <gcli/status.h>
 
-size_t gitlab_get_notifications(gcli_notification **notifications,
-                                int count);
+int gitlab_get_notifications(gcli_notification **notifications, int count);
 
 int gitlab_notification_mark_as_read(char const *id);
